requirements.txt
black==25.9.0
click==8.3.0
iniconfig==2.3.0
mypy_extensions==1.1.0
numpy==2.3.4
packaging==25.0
pandas==2.3.3
pathspec==0.12.1
platformdirs==4.5.0
pluggy==1.6.0
Pygments==2.19.2
pytest==8.4.2
python-dateutil==2.9.0.post0
pytokens==0.2.0
pytz==2025.2
PyYAML==6.0.3
ruff==0.14.1
six==1.17.0
tzdata==2025.2
websocket-client==1.9.0
websockets==15.0.1
